+++ rollup.config.js
@@ -1,7 +1,7 @@
 import { terser } from 'rollup-plugin-terser' // 代码压缩
 import typescript from '@rollup/plugin-typescript'
 export default {
-  input: 'src/index.js',
+  input: 'src/index.ts',
   output: [
     {
       file: 'lib/ja-fetch.js',